refactor: merge controllers
