+++ server/src/http/controllers/admin/eligibleTrainingsForAppointment.controller.ts
