+++ server/src/http/controllers/application.controller.ts
