+++ server/src/http/controllers/applicationAPI.controller.ts
