+++ server/src/http/controllers/formations.controller.ts
@@ -1,8 +1,8 @@
 import { zRoutes } from "shared/index.js"
 
-import { trackApiCall } from "../../../common/utils/sendTrackingEvent.js"
-import { getFormationQuery, getFormationsQuery } from "../../../services/formation.service.js"
-import { Server } from "../../server"
+import { trackApiCall } from "../../common/utils/sendTrackingEvent.js"
+import { getFormationQuery, getFormationsQuery } from "../../services/formation.service.js"
+import { Server } from "../server.js"
 
 const config = {
   rateLimit: {
