--- server/src/http/controllers/metiers/type.ts
@@ -1,5 +1,0 @@
-import { IAppellationsRomes, IMetiers, IMetiersEnrichis } from "../../../services/metiers.service.types"
-
-export type TGetMetiersResponseSuccess = IMetiers
-export type TGetMetiersEnrichisResponseSuccess = IMetiersEnrichis
-export type TGetAppellationsRomesResponseSuccess = IAppellationsRomes
