+++ server/src/http/controllers/metiersDAvenir.controller.ts
@@ -1,7 +1,7 @@
 import { zRoutes } from "shared/index.js"
 
-import { getMetiersDAvenir } from "../../../services/diagoriente.service"
-import { Server } from "../../server"
+import { getMetiersDAvenir } from "../../services/diagoriente.service"
+import { Server } from "../server"
 
 export default (server: Server) => {
   server.get(
