+++ server/src/http/controllers/partners.controller.ts
