--- server/src/http/controllers/shared.types.ts
@@ -1,4 +1,0 @@
-export type TResponseError = {
-  error: boolean
-  message: string
-}
