+++ server/src/http/server.ts
@@ -17,38 +17,37 @@ import { localOrigin } from "@/common/utils/isOriginLocal"
 import config from "../config"
 import { initBrevoWebhooks } from "../services/brevo.service"
 
-import appointmentsController from "./controllers/appointments/appointments.controller"
-import formationsRegionV1Route from "./controllers/formations/formationRegion.controller"
-import formationsV1Route from "./controllers/formations/formations.controller"
-import jobsV1Route from "./controllers/jobs/jobs.controller"
-import jobsEtFormationsV1Route from "./controllers/jobsEtFormations/jobsEtFormations.controller"
-import metiers from "./controllers/metiers/metiers.controller"
-import rome from "./controllers/metiers/rome.controller"
-import metiersDAvenirRoute from "./controllers/metiersdavenir/metiersDAvenir.controller"
+import adminAppointmentRoute from "./controllers/admin/appointment.controller"
+import eligibleTrainingsForAppointmentRoute from "./controllers/admin/eligibleTrainingsForAppointment.controller"
+import adminEtablissementRoute from "./controllers/admin/etablissement.controller"
+import formationsRoute from "./controllers/admin/formations.controller"
+import application from "./controllers/application.controller"
+import applicationAPI from "./controllers/applicationAPI.controller"
+import appointmentRequestRoute from "./controllers/appointmentRequest.controller"
+import campaignWebhook from "./controllers/campaignWebhook.controller"
+import { coreRoutes } from "./controllers/core.controller"
+import emailsRoute from "./controllers/emails.controller"
+import etablissementRoute from "./controllers/etablissement.controller"
+import etablissementsRecruteurRoute from "./controllers/etablissementRecruteur.controller"
+import formationsRegionV1Route from "./controllers/formationRegion.controller"
+import formationsV1Route from "./controllers/formations.controller"
+import formulaireRoute from "./controllers/formulaire.controller"
+import jobsV1Route from "./controllers/jobs.controller"
+import jobsEtFormationsV1Route from "./controllers/jobsEtFormations.controller"
+import login from "./controllers/login.controller"
+import metiers from "./controllers/metiers.controller"
+import metiersDAvenirRoute from "./controllers/metiersDAvenir.controller"
+import optoutRoute from "./controllers/optout.controller"
+import partnersRoute from "./controllers/partners.controller"
+import rome from "./controllers/rome.controller"
+import trainingLinks from "./controllers/trainingLinks.controller"
+import unsubscribeLbaCompany from "./controllers/unsubscribeLbaCompany.controller"
+import updateLbaCompany from "./controllers/updateLbaCompany.controller"
+import userRoute from "./controllers/user.controller"
+import version from "./controllers/version.controller"
 import { auth } from "./middlewares/authMiddleware"
 import { errorMiddleware } from "./middlewares/errorMiddleware"
 import { logMiddleware } from "./middlewares/logMiddleware"
-import adminAppointmentRoute from "./routes/admin/appointment.controller"
-import eligibleTrainingsForAppointmentRoute from "./routes/admin/eligibleTrainingsForAppointment.controller"
-import adminEtablissementRoute from "./routes/admin/etablissement.controller"
-import formationsRoute from "./routes/admin/formations.controller"
-import application from "./routes/application.controller"
-import applicationAPI from "./routes/applicationAPI.controller"
-import appointmentRequestRoute from "./routes/appointmentRequest.controller"
-import emailsRoute from "./routes/auth/emails.controller"
-import login from "./routes/auth/login.controller"
-import campaignWebhook from "./routes/campaignWebhook.controller"
-import { coreRoutes } from "./routes/core.controller"
-import etablissementRoute from "./routes/etablissement.controller"
-import etablissementsRecruteurRoute from "./routes/etablissementRecruteur.controller"
-import formulaireRoute from "./routes/formulaire.controller"
-import optoutRoute from "./routes/optout.controller"
-import partnersRoute from "./routes/partners.controller"
-import trainingLinks from "./routes/trainingLinks.controller"
-import unsubscribeLbaCompany from "./routes/unsubscribeLbaCompany.controller"
-import updateLbaCompany from "./routes/updateLbaCompany.controller"
-import userRoute from "./routes/user.controller"
-import version from "./routes/version.controller"
 import { initSentryFastify } from "./sentry"
 
 export interface Server
@@ -160,8 +159,6 @@ export async function bind(app: Server) {
       appointmentRequestRoute(typedSubApp)
       partnersRoute(typedSubApp)
       emailsRoute(typedSubApp)
-
-      appointmentsController(typedSubApp)
 
       /**
        * LBA-Recruteur
